This is kind of the gist of it. There is the type error I have to fix, but actually despite what I wrote in my earlier commit, something just occurred to me now.
I was not intending it at all, but somehow or the other, what I did is partition the Cuda compiler into pieces.

The map_all_template is in effect a sublanguage for inside of the method body.
The map_module_template is a sublanguage for the outside definitions.

Maybe I'll make more as things go along.

But unlike with DUs, because of this I can suddenly define a method inside a method.

And this is actually very good – if I had implicit parameters I could not impose this restriction! Also I could not do this with DUs without runtime checking. I was worrying how to bring the languages together, but maybe it is fine if I keep them split apart?

Yeah, forget the reader monad, this is it. I was closer to being right the first time.

Viva la template pattern. And amazingly enough, I never really changed the fucking templates at all.

This is quite remarkable, now with this I really do have an unchanging anchor with while to separate specification from implementation.