Solved the type error.
I know that I said that implementation details should not be a part of template, but order of evaluation should be an exception to that.

Unlike before, I am more at peace with what I have written in terms of design. The exception is all this useless paramenter passing. In the cuda_map_module_template I should only really need a single argument.

I am really overdoing it here aren't I?

Let me fix this now.