I've made proper cut for the forward and backward map modules. Now all their particular piees are in perfect order.
With this, I think it is finally time to move to actually implementing the particular map modules.

To start things off, I'll try a few map_1_1 modules.

I'll do it this time. I will pass this Cuda compiler hurdle and then I will have the proper base to finish the rest of the library and the RL project in general.

I'll admit that I want to study Racket more right now; the call of macrology is strong and the type system work in the Turnstile (http://www.ccs.neu.edu/home/stchang/popl2017/) library is quite alluring to me.

But that would be an unberable investment, so I will hold myself back for now. A few years down the line I will dive into this.

I will need it eventually, but it seems that this time I will get bailed out; in terms of workable abstractions, I seem to be at a near optimal rate right now for the tasks I am trying to do.

It can be done with this.

Let me implement those specific modules.