SpiralV5CudaCodeGen.fsx(268,32): error FS0001: Type mismatch. Expecting a     ((string list -> 'a) -> int * string list -> 'a) -> ('b -> unit -> unit * unit) -> ((string list -> 'c) -> int * string list -> 'c) -> ('d -> 'd list) -> (unit -> 'e list) -> ('f -> 'f list) -> 'g but given a     ((string -> string * (string -> string)) -> 'h -> 'i * 'j) -> ((string -> string * string) -> 'k -> 'l * 'm) -> ((string -> string * (string -> string)) -> 'n -> 'o * 'p) -> ('q -> string list) -> ('r -> string list) -> ('s -> string list) -> 't The type 'string list' does not match the type 'string'
I have no fucking idea what the above is supposed to mean.

I've been starting at this type error for nearly an hour, trying to catch a glimpse of insight and it just occured to me that the best way to deal with this is to try an expand the function calls manually.

That will get me on top of things.