Done with the backward module.
Damn, working with tuples is such a pain in the ass in F#. Well, it is not too bad yet.

I move the compile call directly into the template functions, which is a significant improvement to what I had previously. I am getting further away from my ideal of templates being specifications, but who gives a crap. It is still pretty close and if I wanted to I could make the compile function an optional argument.

Enough of that. It is 6:50pm here and I think it is about time I called it a day.

Tomorrow I will expend brain cells on how to make the forward and backward tuple modules more generic

After that I'll complete all the operations that I need and do the signatures.

After that I will be done with the Cuda backend for the time being.

Right now, despite having some code duplication it is still below 400LOC while the previous backend was at 800LOC.

I think it will be a net gain in the end.

I am satisfied with this.

If compilers can be made using this technique, then it should suffice for neural nets as well.

Now if I could only get through this hurdle of constantly having to regenerate the library.

It feels like I am working on this library forever now. I want it to end, so I can move to the next phase. Hopefully this Cuda backend thing will be the last.

It all depends on how roboust the architecture is right now.