Done with a function for both the forward and the backward passes.
Phew.

I need a break.

Well, writting all those map modules won't be too hard, but before I do that, how about I do one last thing and actually create those typesafe launcher functions?

That will be a remarkable step from the previous iteration of this compiler.

With that I'll have everything set in place to merge this with the evaluator without much interference.

No matter what, my vision of the library really was not in line with what I set out to make at the start of December. Doing it this way is the best.

All this functional composition and partial application will be an example for the future me to follow.