I think I am done for the day.
This stuff that I have written so far in SpiralV5CudaCodeGen is A grade material. It is the dividing line separating the me of last year with the me of today.

Even comparing what I written just a month ago, the difference in skill is clear, let alone the me of last year. This gives me confidence that I will finish it this time for good.

Just now, I've completed the launcher function and added the fourth argument.

To be honest, tuples are kind of a pain in the ass, I wish they were more like lists, that would make them a lot more useful, but despite some code duplication caused by this, everything is going smoothly.

Actually, as the last thing today, let me bring back those list functions for the sake of hadmult and only then will I call it a day.