I smoothed out the language greatly, and now expression macros are much more composable than before.
I got rid of the code duplication issues on the F# side in entirety. Well, apart from the lack of generic operations on tuples, but I have enough arguments to satisfy me for the time being.

Now with that...

To be honest, I did not do much today as I was busy with other things, but rather than call it a day here, let me see if I can put the sum in as well.