I am most of the way there with putting the map_redo_map_1_0_1, but I have a type error that I do not feel like digging through right now.
Instead, I'll just call it a day here.

Only a day or two more of this and I'll be able to get back to work on the evaluator and I won't have to touch the Cuda part until I meet my old nemesis - Layer Normalization again.

Damn that thing. It pretty much blew up my whole library apart.

I am not going to let this happen again.

This time my abstractions are absolutely watertight and I intend to get paid for it...at some indeterminate point in the future by people who would probably prefer that I did not write this code in the first place.