Based on visual inspection hadmult seems fine. Ok, phew. I am finally at the point where I was ten days ago when I decided to redo the compiler and can start work on the evaluator.
All in all, it is better than what I had before, but it feels a bit like I am trying too hard. The abstractions are frayed around the edges

The ideal situation for me is that I would be able to write something like:
cuda_map (fun (x: float32) (o: float32) ->
    o <- 1.0f / (1.0f + exp(-x))

And have it all be typechecked and compiled into proper Cuda code.

Actually I had this in the first iteration of the compiler when I used Cuda quotations, but I did not like that solution as I was parsing quotations into DUs and then operating on them.

It was actually less safe than even the last compiler and felt incredibly hacky.

F# quotations seemed better until I tried using them.

In review, I did do a heroic job here, but having a Cuda DSL this is not. The amount of code duplication in the Cuda code generator is quite frankly ridiculous.

At first I thought I could approach 90% of optimal, but this is more like 40%.

I am decently sure that I could do a far better job in Racket and the next version of the Spiral library is going to be in that language.

I won't work on the evaluator today, instead I will spend the rest of it watching Racket tutorials.

Well, that having said, this is quite good work regardless.

I am convinced that given the language features provided to me, it is probably impossible to do a better job that this.

And that is quite high praise. I can hold my head high here.

To improve on this, one would need either dependent types or macros. And since dependent types are too risky, that leaves only macros.

Even a Phd would not be enough to make me confident of using dependent types.

Enough.

Making the library its own language can wait for the future. My current priority is using what I have now to claim my stake, I've wandered way too much last year, and redoing from scratch in a different language would take too much time.

The recent Predictron paper by Deepmind shows that they are getting serious as well. It is a variation on the idea I was intending to use to conquer poker, so I can't really afford to fool around too much.

...Well, that is actually the wrong way.

I have plenty of headroom to do this properly.

I do not owe it to anyone to start making money in the next six months. The investments into knowledge I am making now are important.

So let me learn some macrology.