Finished those Racket videos. In some sense the last 10 minutes of the second DSL video(https://www.youtube.com/watch?v=JC3o87p_rBY) were the most interesting since Matthew Flatt went into how type checking is done in Racket, I could have just skipped the rest as they were pretty boring.
As usual, when I get in a researchy mood like this, I tend to waste humongous amounts of time just dithering around.

This time I am going to use my foresight a little and see where the the future is going.

And the vision that I see is that it is simply too early.

Lisp existed since the 60s, but this new `Type Systems as Macros` (https://www.youtube.com/watch?v=j5Hauz6cewM) thing is genuinely novel. I've never seen composability on the level of language's type system, and I wanted to see more material in Racket in this direction, but it is incredibly sparse. Typed Racket seems like an afterthought almost.

When I studied macros, both now and in the past, what I really wanted was an aid that would help reasoning at the type level, but all I am getting are syntactic abstractions, which while not useless, are not what I want.

It is simply too early. Racket's real promise is in bridging syntactic and type abstraction using macros.

It won't happen anytime soon.

So where does that leave me?

To start things off - I will backtrack on what I wrote in yesterday's commit. The next version of the Spiral library will be in F# as usual.

Hopefully by then, F# will have implicits (for cleaner DSL syntax) and GADTs (so I can create a well typed language for my RL agents.)

But my the hunch I had at the end of last year when I internalized the template pattern that I have everything I need was in fact correct.

To be honest, I am tired of this library which why it is going so slowly. This mostly a result of working so long on the same damn pieces.

So it will be done slowly, and will be finished slowly - but it will be finished in the end.

Today my mission will be to integrate the new Cuda DSL with the primitives segment.

It occured to me that I could make the Cuda compiler well typed by adding some custom datatypes and defining operators on them, and that I could get rid of those argument size lambdas by making them implicit using statically resolved type params, but to be honest, what I have now is perfectly enough.

The real test of the new Cuda DSL will be when I get to layer norm, which killed my momentum 3 months ago. I can hardly believe it has been that long - this is ridiculous.

I'll keep to my vow of making January more productive that last six month combined.

And the way to do this is not to be faster, or to work harder, but to just do it properly.

I have all the pieces that I need to do it now.