Apart from hadmult, I am now on feature parity with the previous iteration of the library.
To be honest, tuple arguments are such a pain in the ass that I am considering removing them from the library in their entirety.

I can't be assed to write an explicit map for every damn case. So for that sake, I'll do some research on FsControl and see whether there is anything in there to help me a little here.

Actually, generic operations on dependent lists are one of the use cases of GADTs. The depth of the list can also be determined statically as well with GADTs similar to inductive types.

Tuples are a bit like heterogeneous lists if you think about it, but in other cases they fall short.