Oh, oh crap.
If I want something like a map_redo_map with int and a float output, I am going to have to go as far as making a new template, a new backward function and a new b_arg. Basically, new everything for a somewhat trivial change.

My code might be well separated now in terms of concerns, but it is not generic at all. It is much less than I thought it would be.

Oh crap.

If only the dynamic approach had a better syntax.

You know, maybe I should give Racket a chance when I am done with all of this.

For the first time, F# is making me die a little inside.

The irony of this all is that I am not wrong on anything at all with regards to the template pattern. It really is worth that much. The structure of the library that is in my mind is rock solid, it is just that I won't be able to exploit it like this.

No, actually I have a great idea how to salvage all of this. I'll have to make use of custom types and statically resolved type params, but it will work.

I'll do it tomorrow.

I am just too invested in this damn library to give up now, to give up just like this.

I'll simplify the templates drastically and get close to my vision of having specification determine the implementation.

I'll get rid of those consts and cleanly and absolutely separate out the inputs and the outputs.

I'll have to give up on the idea of having input_adjoints be in the last position during the backwards phase, but it is fine.

...

...

(1h later pacing back and forth) No, the above won't work. Right here and right now, I've pushed the language to its breaking point.

One hand I was right about the Template pattern being the true style and the basis for everything else, but what I was wrong on is that F# won't let me take advantage of it.

Moving back to Dus is the wrong answer, but that custom type nonsense with statically resolved type params won't work either.

At this point basically I need to move to two phase compilation.

I need to generate those map modules programmatically, were similarly to how I am generating those Cuda modules.

So I am considering 3 options:

1) Check out Rascal.
2) Find out if Racket's macro language can be used to generate F# code instead of just Racket code.
3) Template pattern + F# + type providers maybe.

My troubles just never end do they?