What a restless night.
I am going to have to try doing it in Racket after all.

This is pretty insane. Of all fucking things, the interface to the damn map module and not the map module itself is what I am finding impossible to do.

Preposterous.

Obviously, I do not relish doing this in Racket as I do not really know its macro system, nor does it have a Cuda bindings, meaning I would have to write them myself.

And then there is everything else I would need to redo in Racket.

But the solution of generating the API programmatically, while it would work is such a hacky solution.

But what I am doing is already hacky.

This is like the third fucking time I did the compiler, it should be obvious to anyone that maybe F# is not the right tool for the job given how much difficulty I am having integrating the Cuda parts with the rest of the language.

And this decision would mean having to spend months (again) rebuilding things from the ground up in addition to learning the Racket's macro system.

Is this worth the investment?

Well, as they say, first you get the abstraction and then you get the control.

I'll take a 'break' from the F# library for a few months.

Just like my story Simulacrum, it seems the Spiral project as well will have to go off the rails and into the abyss before it emerges back out again.

Again, I won't be able to fulfil the promise to myself of making January strictly productive, nor will I be able to report a positive outcome on the RL front in six months.

But...I'll win someday.

What I want is really stupidly simple. In F# one can write Array.map (fun x -> x * x). I just want to be able to program in such a natural functional way with Cuda in the language of my chosing.

If I generate the Api programmatically, I'll be restricted to having to use a different functions for each different combination of arguments.

It is ridiculous. A programmer is only as good as the abstractions he has.

I cannot accept this.