http://eiriktsarpalis.github.io/typeshape/#/
Eirik Tsarpalis even mentions GADTs in his slides. Yeah, this is exactly the kind of problem I could deal with if I had GADTs.

So maybe there is hope for me yet.

Edit: I meant to push this yesterday.

http://stackoverflow.com/questions/7213599/generic-higher-order-function/7221589#7221589

This is the post Erik linked to me explaining why he uses interfaces instead of functions. I am going to have to put the library development on hold until I understand all of this higher ranked type magic.

Ugh, this interface hopping is quite nasty, but it is either that or moving to Racket. The techniques used here remind me a little of the stuff I read about immitating GADTs in C# paper.