I've gone through the TypeShape's source.
I am absolutely certain right now...TypeShape can bail me out here and give me lifeline on GPU programming enough so that I do something useful with the Spiral library.

But the feeling of dread I now have is not wrong either. And neither was much hunch in December that the dynamic style in F# is the most flexible. And in my blog update, I was enthused about functions, but in the end, I can only do so much with them.

Going through TypeShape's source convinced me of this single fact – if I want to do GPU programming properly, I absolutely need more syntactic abstraction. I need the absolute expressive power afforded by macros.

TypeShape is truly an awesome library, but it won't give me the basis for a language.

With this, I won't have to get rid of the library entirely, but I'll only be able to make Spiral the library and not Spiral the language.

Whatever the case is, it seems I am going to be putting some hours into learning macrology either way.

Maybe as a compromise, I could put some time into learning IronScheme or ClojureCLR?

Right now I'll call it a day, but as for tomorrow…

I'll think whether I want to work on Spiral or learn some Lisp.