This is such a complicated situation for me. Let me gather my thoughts.
1) With TypeShape (or reflection in general) I could write the appropriate interface for the Cuda modules.

The issue - this is such a hack if there ever was one. If I was going to make a language and ended up writting code like this, I'd have serious doubts about the endeavor.

Which I do have right now. I could do it like this, but I am pretty much at existential crisis stage right now because of how the code would come out.

I am not getting paid to write the library, so the real payment I get is how good the code I am writting is. It is ironic that it would come to this after figuring out the template pattern and writting the best code of my life.

Not being able to write the correct abstractions fills me with a feeling of poverty.

2) Haskell could do this. Typeclasses support return type polymorphism, so all I have to do is give them a type and they would do the rest.

That roboust printer in TypeShape could be done using typeclasses, as could the other examples except with 100 times more elegance.

Edit: Well, to be fair, I am not sure about maping over tuples in a polymorphic way.

3) I do not know exactly how, but Ocaml with GADTs could do this as well. Haskell supports GADTs as well.

The trouble with going with Haskell though, is that DL algorithms are mutable by nature and this would become a problem.

I'd be able to do the Cuda thing as it should be done, but like F# now, I'd get fucked when it is time to get serious with the library.

Ocaml on the other hand has pretty poor Windows support.

It is such a tragedy that Haskell is pure and lazy, otherwise it really woudl be the best language in existence. Typeclasses and GADTs are othogonal to the purity and laziness thing.

5) In December I actually went with a dynamic approach to the library, but changed my mind because all the passing of union types made it exceedingly tedious.

If the syntax for this was better, I'd probably go with this in F#, type safety be damned.

Seriously, fuck these Cuda modules. I can't believe that making the damn compiler was much easier than actually integrating it with the rest of the langauge.

I had no expected this at all - it is so ridiculous.

Ok...for today, how about I just hold my nose, bring in TypeShape, do the interfaces for the modules properly, and pretend I was not the one who really wrote that code?

That seems like the way to do it.

And as a safety measure, I'll put in some time into Racket, so I do not get bushwhacked by things like this again.

It seem no matter which language you pick, there is never really a perfect fit.