Ok, I've decided.
let mkCudaLambdaArgs<'ins, 'consts,'outs>(f: 'ins -> 'consts -> 'outs) : 'ins * 'consts * 'outs =
    Unchecked.defaultof<'ins>,Unchecked.defaultof<'consts>,Unchecked.defaultof<'outs>

I'll use the `Unchecked.defaultof` trick to get around this hump.

With that decided there are still some lingering issues to think about. I've been thinking hard for the last few days, obsessing in fact over the matters of style.

I still am 100% certain that the template pattern which I discovered at before the start of this year is the way to program. Functions should be passed to functions and like those modules, they should form isolated domains.

This is essentially the future of programming

What has taken me off guard is the need for metafunctions that produce values based on types.

I had not seen this coming and I never anticipated this become a part of my new purified style. But I suppose if I thought more deeply about it, it should have been occured to me that at some functions would in fact need the information to compile themselves.

So the reason these map modules are giving me trouble because in essence, I stumpled upon the need to describe self compiling functions.

Quite a novel concept.

The other problem that won't be a problem really for these modules here, but is in general is that the functions I am now using, or rather the template pattern style will run into problems if I start passing around values.

For example if I now add CudaInt and CudaFloat and then add arrays for them and const variants. Then suppose I want to add the address function to return their address - well if I passed that address function in a lambda, I would need to make a separate function for each variant of the data structure!

So it would have to be at the top level, but I am starting to suspect that the correct way to program is to pass functions around and put less and less stuff at the top level. For the same reason global variables are wrong, so are top level functions.

I've felt it very vividly, how easy it is to go straight for the specification when you do not reach for the top level function as the first thing.

I'll admit, even though F# could do more to help me in this, I was wrong for not anticipating the need for metafunctions, but in general this reveals the weakness of the langauge - the type system is actually too weak to let write the way I want.

But I had not resolved myself to program in a dynamic language, and I fully belive that type inference is like a brainjack into the compiler's mind.

Right now, maybe there is in fact a very real danger that I am outgrowing it. I seem to be bashing my head quite a bit against the compiler lately.

I need some time to play around. Maybe I do not even need macros in the isolated cases, but just more significant function passing and the ability to define metafunctions.