No, even though I figured out how to disassemble the type, it is no good.
Sure I could introduce CudaInt and CudaFloat, but then I would have to redefine the operators for literally every kind of expression.

Ok, I give up.

I'll go with the metaprogramming approach in the end. I'll just make a function to print out the map modules. That will be well typed and quite reusable.

It is not that I will use that many different types of modules.

I am too tired to deal with this crap.

I tried Racket seriously for a bit today, but while the idea of being able to do anything sounds good on paper, that lack of any contextual cues is really putting me off. I worked on the Knapsack problem for a bit and when I got back, I lost the thread completely.

Racket code looks just like a bunch of text to me.

I guess I should be pressing Ctrl + R a lot more when programming in Racket, but in F#, I am used to programming without running the program for days.

I guess having GADTs here would really make a difference. It would cleanly separate the printing from type checking, while like this, they would be pretty much bound together had I decided to introduce well typed terms. This is the reason it is not worth doing.

And as an alternative, if I had to convert CudaInt or CudaFloat to string manually, then what is the point of this in the first place?

I'll make up for this horrendous loss of time in the last week by putting some time into Haskell...which would be more wasted time, but maybe it will help me get over my regret.

At the very least, with the metaprogramming approach, I definitely am operating at the right level of abstraction even if the convenience won't be quite there.