Done with the forward meta functions. I need a break from this.
This is real drudge work, but it beats writting it by hand.

It is unfortunate, but the only two mainstream languages that have GADTs are Ocaml and Haskell. In the following days, I am going to see if I could dig up some info on how deal with mutable structures in Haskell.

I am aware that the way I approached Haskell was not exactly optimal. To a degree I should have let is sink in slowly, instead of trying to break the language all at once.

But still, I feel as if I am just chosing from bad options.

Maybe the correct option would have been to write the library in D which is not dynamic and yet has good metaprogramming support. Also Ocaml has gotten new macros as well.

...Forget it. You do your best and then you do whatever it takes to win.

I want to take a break, but let me see if I can finish all the modules now. The first one was the most difficult. From here, I should be able to reuse some code.

The actual generated interfaces are quite simple now.