I'll stop for the day here. I'll finish the backward tomorrow.
It feels like this is taking a lot, but in fact it is taking a lot less that it was for the first batch of interfaces.

Given that it almost seem magical that it took me this long to just stop and start writting this stuff.

The irony is that I kind of expected this, but still had to go through the motions because style matters.