No...
The metaprogramming approach was the wrong choice.

I am going to backpedal on this decision and give up on static safety for the sake of expressiveness once again.

Or rather, I will go back to v5 and if I require ints in the output for map_redo_map for example, I will in fact make a special case just for that.

I thought it would be simpler, but just this code to create the interfaces is already on par with the modules in size, and not too mention because I am passing around strings everywhere, I do not have any access to type information to guide me.

Even though I have written it, I can't keep track of this code. And since I am not getting paid for it, I might as well throw it away as it is no good to me.

My heart is in shambles over this. This is true agony.