It seems I really will be spending the following time studying Haskell. I have no choice as working on this library is giving me no satisfaction, I've just been bashing my head against the wall.
Since the DO course is running again, I've just been working on the Knapsack assignment.

The focus this time is on figuring out the IO monad. One thing that occured to me now that was not obvious to me a year and a half ago is that that for the larger problems, the way trace is build is quite similar to what I am using in the library for the backprop algorithm (and for Q learning.)

It seems I will be working on this problem for the entire day today, but I will at least try to check out Ocaml as well as an alternative.

I've bashed Haskell for its poor IDE support, but Ocaml as far as I can tell does not even have it, and the Windows version of it is 2 updates behind the Linux one.

That is why Ocaml was not even a consideration.

It will be quite humiliating to go to Haskell after all this time in F# and after the way I praised function in the last Simulacrum update, but I have no choice.

Machine learning, reinforcement learning, the Spiral library, money, friends, family...none of that matters anymore.

What matters is that I finish the map module.

Functional programming is great - just compare it to Java a while ago where you had to declare a class and 30 lines of boilerplate just for a simple lambda.

This is the situation I have right now in my own library and it is unbearable.

I tried Haskell before, but that was was where I was trying to balance various constraints.

Now my determination is at an entirely different level.

I will stop at nothing and I will sacrifice anything to complete the map modules to a satisfactory level. It will be done no matter what. Absolutely.

I'll hold off on updates for a few days in this repo until I get the ball rolling. Cyao.