This winter is such a killer. I keep getting old people's diseases. Yesterday I had to lay low for an entire day because I could not use my hands.
It gave me time to think.

I do too much of that, but before I defect to Haskell for real instead of just flirting with betrayal, let me try the tagless idea.

The Template pattern was a glimse into the future of programming to me, but unfortunately, no functional programming including Haskell can accomodate such a style. GADTs+type families come close maybe.

But if I wanted to do this properly, Cuda programming should really be a language extension or an outside DSL.

The only bright side in all of this is that I won't have to do much work with regards to loop fusion as backprop requires all of the intermediates.

So let me do the typeclasses with statically resolved params (F# style) and I'll reverse the template pattern to be used with shallower hierarchies.