http://stackoverflow.com/questions/41813454/how-to-resolve-the-strange-type-error-in-a-recursive-map-with-statically-resolve/
I have to offer my thanks to Gustavo, but I am now 95% sure that I will be switching languages. The tagless style cannot be immitated in F# due to compiler bugs.

Edit: I can't belive he managed to find a way around this. That guy is a genius. I guess I'll stick with F# for a while longer.

Seriously, if Haskell was impure and strict like Ocaml and F#, I would have been using it from the start.

That having said, let me see if I can implement the variables.