Another type erorr that I can't figure out. This would be enough to trigger my `blame the language` rule.
Yeah, statically resolved type params aren't typeclasses. They could be enough in theory, but it is safe to say that I was 100% wrong on this. The tagless style is even more difficult to emulate than what I'd imagined in F#.

Yeah, the last 3 weeks were despressing as hell for me.

Should I pack up my bags and move? Or should I do some kind of after action report as per Gustavo's suggestion?

I've been thinking about doing a shallow DSL, but that would be the same as the second Cuda compiler.

....Yeah, I do not feel like digging through the past commits. I'll let the sleeping dogs lie. Today and yesterday I pretty much proved that F# lacks the necessary type system features to let me implement what i want.

Between Haskell and Ocaml, with Haskell, I just might be able to take advantage of the Cuda bindings for it.

But with Ocaml, I might be able to adapt whatever Linux bindings it has for Windows.

But Haskell's laziness aside, I never want to go through the pain of having the language I am using insufficient to implement my vision.

In that regard Haskell is in a league of its own.

There is no choice - I think I will give it a serious try.