Ok, now that I am a Haskeller and diving into the language with intention to finish it, I had the chance to bring the tagless printer closer to the GADT one.
The thing that did not occur to me while reading the tagless paper is that instead creating special classes for expressions, I could just create fresh Num, Eq and Ord instances...

Which while an interesting thought, would mean that I cuold not abstract over both the printer and the evaluator.

Er, nevermind.