It seems that thanks to the RebindableSyntax extension, Haskell's syntax is more maleable than I thought. This is actually a great news as far as my Cuda DSL is concerned. Syntax was kind of a pain point in F# as well, so I am glad I found this.
At any rate, I am done with the evaluator for the expressions part of the DSL is done. For today I had enough. It took me almost 3 weeks to defect to Haskell, and it has been quite draining for me emotionally. I will have to leave so much behind, but now that I have the skill, maybe I have what it takes to rebuild much faster than I think.

Depending on the state of the Cuda package, I might have to make my own bindings for the library. This will be a pain in the ass - ManagedCuda was one of the reasons why I kept delaying the decision to make the move.

But with Haskell, I will do absolutely everything right this time. Absolutely.

No doubt that in terms of language features, the thing is unparalleled. I won't have to agnoze about expressing things in it.

And rather than use my own DSL, I might end up piggybacking of Accelerate for the map operations.

The most important is that move on from the conflict raging inside of me that I had suffer for the past three weeks because insufficient language features of F#.

Peace of mind is something that money can't buy, but Haskell might give me.

A programming is only good as his abstractions.

Tomorrow, I will grip the reigns on my will and ride onwards.

I'll break out the state monad and start work on the statements part of the DSL.