Ok, I've been studying the Tagless-Staged library I found on Oleg's site and I can't make heads or tails of it. It is extremely academic.
The type level composition was interesting for me to know that it is possible.

I'll remove it in the next commit.

The thing I'll look at next is the Accelerate library.

I've only dabbled in it before, but now I am rather interested how it handles variable names and such, more than the optimizations underneath.

It probably has a lot worth studying for me. Also, I still feel like I have some obligation to contribute to Futhark as well. Maybe that I am now not using Haskell in 'anger' so much, I'll be able to make better progress.

Both Accelerate and Futhark are worth studying, but Accelerate more so as it is a GPU language integrated with Haskell. I am definitely interested in the integration aspects of it now.