Ok, I've gone through the Accelerate's source for the second time in my life to see if it makes more sense this time. It does not.
Large languages like Futhark and Accelerate seem so monolothic to me from the outside.

And it does not seem like Accelerate uses shared memory.

Futhark does, but it is likely to be brittle and I am skeptical of how well it handles memory given that the author had to ask me for allocation speeds. But that is not the problem.

The actual problem is that I am skeptical of the general architecture of these languages.

Both Scala and Ocaml and Haskell are smart, but their optimizations break all the time. Scala also has that infamous loop issue that has been open for ages, that Oleg managed to solve recently in his paper.

https://arxiv.org/abs/1612.06668
Stream Fusion, to Completeness

It reinforced the bias that I had that to get around these performance issues, smart macro use will be necessary rather than functional composition. The above is a dumb approach, but clearly effective, while constantly shoving optimizations into the compiler is a dead end as the state space is too big.

Generating C++ code from Haskell is a bit like making Haskell a macro system for C++. It is a true test for a language - how well it handles deep hierarchies. The template pattern I had a glimpse of is undoubtedly the future of programming, but there is no language that can make use of it, save for dynamic ones.

Ahh, I have no idea.

I am simply going to have to give it a show. Bring out the monads, and make Haskell my super-Lisp. I have no idea how it will turn out until I try.

To be honest, I do not like how Accelerate's source looks, I can't really find a clear separation between specification and implementation.

Ah, I change my mind on advanced type systems. Type inference is the soul of a programing language and language features its bones.

I should make this my goal - while I experiment with the Spiral library in Haskell, what I should really be checking out are realistic variants of the template pattern. F# does not really have what it takes to carry that idea very far. I might get further in Haskell.

At any rate, I do not have any choice in the matter. If I want to get better, this is the path that I have to travel.