Right now I am gripping with typeclass related type errors right now and it is getting late, so let me push this here. I had a question for SO, so I was intending to finish this part, but I got bogged down. To be honest, I am not sure how to deal with this part. I think maybe I should just make the damn monad instance for CudaOuter that returns all the statements in a list.
I originally intended to use the writer monad here, but that thing is useless.

Yeah, I am getting a bit lost in types here. But the annoying thing here is that now Haskell now has Monad as a superclass of applicative and functor so I do not feel like implementing it.

Maybe I can just make it undefined?