To be honest, I am thinking of throwing in the towel on abstracting the interface.
I spent the last two and a half days going through the Haskell User Guide and now have a decent understanding of all of its extensions. To be honest, I like typeclasses and all their extensions, HKTs included. Counting all those extensions as one, I would go as far as saying that typeclasses are the single greatest feature of any language that exists.

But I am coming the conclusion that the reason why they exist in the first place is because the language lacks proper impredicative polymorphism which would allow it to generalize higher order functions.

The template pattern was not a mirage and neither was I on the wrong track - the languages (both F# and Haskell) are wrong. I've seen the future and it looks pretty much like the past.

Furthermore, the whole reason why I am trying to abstract the damn C++ code is performance and functional languages in general have performance issues with abstractions - the way out is staging and heavier use of macros.

So basically Racket has all the elements in place, but they certainly aren't fitting together.

For the past 1.5 years, I've been pitting the languages against each other in my mind, but I am not getting any satisfaction out of fermenting that kind of conflict anymore.

Maybe the way to upskill myself would be write my own typechecker and my own macros and abstract the C++ code that way.

The sickening thing with moving to Haskell, is that apart from this Cuda piece, I think that the rest of the code that I have written in F# is perfectly fine, and if I had to rewrite it in Haskell, I would be dropping performance all over the place.

The most sickening thing about laziness in Haskell is that ML library actually uses laziness in places for performance - if I turned on the strictness pragma like I did for the knapsack, I would not be able to put the laziness back in selectively. The reason being is that the type system is not tracking it!

Haskell just has such a different set of values than do ML languages.

It is true that typeclasses + extras are orthogonal features to the whole purity and laziness deal, and F# should definitely take them from Haskell, but I can't really afford to wait for this to happen even if it is inevitable.

It is true that the interface is not that important...

No scratch that, it is important, I could have always used quotations as well too, but I just did not feel like doing typechecking myself.

Erm, let me try combining the quotations and DU idea, and the template pattern idea. Right now I have three different Cuda compilers for the Spiral library (this Haskell stump being excluded). Maybe I just need to merge the things together in order to get something good?

...I definitely want to reach beyond the possible. Those shitty tuples will still annoy me even with typeclasses and I really should not have failed at that textual inteface generator.

I'll do this as long as it takes, but today I do not feel like it anymore. To be honest I barely did anything other than finish the guide.

http://strictlypositive.org/Easy.pdf
Simply Easy!
An Implementation of a Dependently Typed Lambda Calculus

Yeah, I want to study this for a bit. To be honest, if I was doing it raw myself, maybe dependently typed programing would be easier.

I might throw in the towel on the interface, but I won't go down that easily without extending my reach.